Use Serilog.Sinks.Console instead of Serilog.Sinks.Literate
The features of Serilog.Sinks.Literate have now been merged into the default Serilog console sink.

The motivation behind this change is to be able to easily change the theme `WriteTo.Console(theme: ConsoleTheme.None)` so that the output in the JetBrains Rider console isn't white text on white background.