Display error panel instead of silently logging
Pull request #89 changed error handling from alert() to log() because using window.alert for errors stops the refresh on monitoring displays.

Unfortunately, this is a nightmare from a user experience point of view on non-monitoring displays because in case of error, the "Loading…" screen stays forever.

This pull request addresses this issue by displaying an error panel in case of error.