Convert solution to the new csproj format
Although the [documentation (Additions to the csproj format for .NET Core)][1] seems to imply that this is a .NET Core feature, it works with .NET Framework too and is much easier to maintain.

With this commit, the OctoPack feature is gone because the new .csproj format is not yet supported by OctoPack. Here is the [suggested workaround][2]:
> We have this support on our radar, but don't have an ETA at this stage.
In the meantime, you could try running `dotnet publish` and then run `octo.exe pack` against the publish output folder.

[1]: https://docs.microsoft.com/en-us/dotnet/core/tools/csproj
[2]: https://github.com/OctopusDeploy/Issues/issues/3822#issuecomment-333706392

# Conflicts:
#	src/SbManager/SbManager.csproj
#	src/SbManager/packages.config